im done, fixed ignore

+++ shell/install_packages.sh
@@ -4,7 +4,6 @@ log_file="$(dirname "$0")/install_progress_log.txt"
 
 sudo pacman -Syuq --noconfirm
 
-
 download_package()  {
     sudo pacman -S --noconfirm $1
     if type -p $1 > /dev/null; then
@@ -23,11 +22,11 @@ echo "Trying to install command-line utils..." >> $log_file
 
 download_package tmux
 download_package ranger
-download_package sl
 download_package ripgrep
 download_package unzip
 download_package wget
 download_package curl
+download_package sl
 
 echo >> $log_file
 
@@ -35,27 +34,33 @@ echo >> $log_file
 # --------------
 # Xorg and GUI
 # --------------
-echo "Trying to install GUI..." >> $log_file
 
-download_package xorg
-download_package xorg-xinit
-download_package xwallpaper
-download_package brightnessctl
-download_package sxhkd
-download_package pamixer
+read -n 1 -p "Install graphical environment? [y/n]" prompt
+if [[ $prompt =~ [yY] ]] then
+    echo "Trying to install GUI..." >> $log_file
 
-download_package mpv
-download_package zathura
-sudo pacman -Sq --noconfirm zathura-pdf-mupdf
+    download_package xorg
+    download_package xorg-xinit
+    download_package xwallpaper
+    download_package brightnessctl
+    download_package sxhkd
+    download_package pamixer
 
-echo "Trying to fonts..." >> $log_file
-sudo pacman -Sq --noconfirm ttf-jetbrains-mono
-sudo pacman -Sq --noconfirm ttf-jetbrains-mono-nerd
-sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols
-sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols-mono
-sudo pacman -Sq --noconfirm noto-fonts-cjk
+    download_package mpv
+    download_package zathura
+    sudo pacman -Sq --noconfirm zathura-pdf-mupdf
 
-echo >> $log_file
+    download_package xournalpp
+
+    echo "Trying fonts..." >> $log_file
+    sudo pacman -Sq --noconfirm ttf-jetbrains-mono
+    sudo pacman -Sq --noconfirm ttf-jetbrains-mono-nerd
+    sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols
+    sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols-mono
+    sudo pacman -Sq --noconfirm noto-fonts-cjk
+
+    echo >> $log_file
+fi
 
 
 # --------------
